+++ src/main/java/com/dmmsoft/app/analyzer/analyses/revenue/InvestmentRevenueResult.java
@@ -3,23 +3,25 @@ package com.dmmsoft.app.analyzer.analyses.revenue;
 import com.dmmsoft.app.analyzer.analyses.AnalysisCriteria;
 import com.dmmsoft.app.analyzer.analyses.AnalysisResult;
 
+import java.math.BigDecimal;
+
 /**
  * Created by milo on 15.04.17.
  */
 public class InvestmentRevenueResult extends AnalysisResult {
 
-  private  double capitalRevenueValue;
-  private  double capitalRevenueDeltaPrecentValue;
+  private BigDecimal capitalRevenueValue;
+  private BigDecimal capitalRevenueDeltaPrecentValue;
 
-    public double getCapitalRevenueValue() {
+    public BigDecimal getCapitalRevenueValue() {
         return capitalRevenueValue;
     }
 
-    public double getCapitalRevenueDeltaPrecentValue() {
+    public BigDecimal getCapitalRevenueDeltaPrecentValue() {
         return capitalRevenueDeltaPrecentValue;
     }
 
-    public InvestmentRevenueResult(double capitalRevenueValue, double capitalRevenueDeltaPrecentValue, AnalysisCriteria evaluatedInput) {
+    public InvestmentRevenueResult(BigDecimal capitalRevenueValue, BigDecimal capitalRevenueDeltaPrecentValue, AnalysisCriteria evaluatedInput) {
         this.capitalRevenueValue = capitalRevenueValue;
         this.capitalRevenueDeltaPrecentValue = capitalRevenueDeltaPrecentValue;
         super.finallyEvaluatedInput = evaluatedInput;
